requirements.txt
python-consul==0.7.0
requests==2.13.0
six==1.10.0
